Rename "ConfirmedCasesAttributeCard" to "PatientsDetailCard"
