+++ components/cards/PatientsDetailCard.vue
